+++ CHANGELOG.md
@@ -9,7 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.6.2] 2022-05-12
 
+### Added
 
+### Changed
+
+- Remove redundant default and project file paths mountings.
+
+### Fixed
 
 ## [0.6.1] 2022-05-12